Fix org error on comment
Something in the check for whether point is in a source block or not
was throwing errors for non-org buffers. Without digging into the
source of org-in-src-block-p, it seems like something in there expects
that function to only be called within org buffers.  Now just check
first if it's an org buffer.